# Initial topdown optimizations