--- a/storage/dump.go
+++ b/storage/dump.go
@@ -14,3 +14,13 @@ func Dump(ds *DataStore, w io.Writer) error {
 	e := json.NewEncoder(w)
 	return e.Encode(ds.data)
 }
+
+// Load reads the content of a serialized DataStore from the io.Reader r.
+func Load(r io.Reader) (*DataStore, error) {
+	d := json.NewDecoder(r)
+	var data map[string]interface{}
+	if err := d.Decode(&data); err != nil {
+		return nil, err
+	}
+	return NewDataStoreFromJSONObject(data), nil
+}