--- a/topdown/aggregates.go
+++ b/topdown/aggregates.go
@@ -55,7 +55,11 @@ func reduceSum(x interface{}) (ast.Value, error) {
 	if s, ok := x.([]interface{}); ok {
 		sum := ast.Number(0)
 		for _, x := range s {
-			sum += ast.Number(x.(float64))
+			f, ok := x.(float64)
+			if !ok {
+				return nil, fmt.Errorf("sum: input array contains non-number value")
+			}
+			sum += ast.Number(f)
 		}
 		return sum, nil
 	}