This PR contains a few optimizations that I made to the topdown evaluation last week after doing some profiling. The individual optimization commits are fairly small and isolated.

Here's the output of running the scheduler benchmark on the last 6 commits:

```
[work](benchmark)torin:~/Development/Styra/work/src/github.com/open-policy-agent/opa$ for rev in $(git rev-list HEAD...52b6684a8c03efe51cd4e166b8b60e5197cb747d | head -n 10 | xargs); do git checkout $rev; make perf; done
Note: checking out '533784940a1e1304fa9e6db1e780cd4e2fda690c'.

You are in 'detached HEAD' state. You can look around, make experimental
changes and commit them, and you can discard any commits you make in this
state without impacting any branches by performing another checkout.

If you want to create a new branch to retain commits you create, you may
do so (now or later) by using -b with the checkout command again. Example:

  git checkout -b <new-branch-name>

HEAD is now at 5337849... Optimization: Avoid slice copies in hotspots
go generate
go test -v -bench=. ./test/perf/.../
testing: warning: no tests to run
PASS
BenchmarkScheduler10Nodes-8	       1	2803496994 ns/op
ok  	github.com/open-policy-agent/opa/test/perf/scheduler	2.854s
Previous HEAD position was 5337849... Optimization: Avoid slice copies in hotspots
HEAD is now at f106040... Optimization: Avoid plug on indexing check
go generate
go test -v -bench=. ./test/perf/.../
testing: warning: no tests to run
PASS
BenchmarkScheduler10Nodes-8	       1	3111978150 ns/op
ok  	github.com/open-policy-agent/opa/test/perf/scheduler	3.158s
Previous HEAD position was f106040... Optimization: Avoid plug on indexing check
HEAD is now at b840758... Optimization: Remove copying of bindings
go generate
go test -v -bench=. ./test/perf/.../
testing: warning: no tests to run
PASS
BenchmarkScheduler10Nodes-8	       1	3258621573 ns/op
ok  	github.com/open-policy-agent/opa/test/perf/scheduler	3.305s
Previous HEAD position was b840758... Optimization: Remove copying of bindings
HEAD is now at 30e7066... Optimization: Defer unification until needed
go generate
go test -v -bench=. ./test/perf/.../
testing: warning: no tests to run
PASS
BenchmarkScheduler10Nodes-8	       1	3757969676 ns/op
ok  	github.com/open-policy-agent/opa/test/perf/scheduler	3.804s
Previous HEAD position was 30e7066... Optimization: Defer unification until needed
HEAD is now at d0cd968... Optimization: Improve string and var term hashing
go generate
go test -v -bench=. ./test/perf/.../
testing: warning: no tests to run
PASS
BenchmarkScheduler10Nodes-8	       1	3994262904 ns/op
ok  	github.com/open-policy-agent/opa/test/perf/scheduler	4.041s
Previous HEAD position was d0cd968... Optimization: Improve string and var term hashing
HEAD is now at d12d3a4... Vendoring: Add siphash and update dependencies
go generate
go test -v -bench=. ./test/perf/.../
testing: warning: no tests to run
PASS
BenchmarkScheduler10Nodes-8	       1	4915227223 ns/op
ok  	github.com/open-policy-agent/opa/test/perf/scheduler	4.962s
[work](detached*)torin:~/Development/Styra/work/src/github.com/open-policy-agent/opa$
```