@timothyhinrichs 2016-07-11
Github isn't letting me add this comment inline for some reason. It pertains to the Binding algorithm...

Trying to think through if this always gives us the right answers.

In this scheme we're (i) changing the existing dictionary and (ii) allocating a new struct to record the old value.  Every unbind changes the existing dictionary back to what it was.  Does every plug look up a value in a dictionary, or does it transitively close the individual bindings?  I'm wondering about an example like a sequence of bindings x:y, y:z, z:1.  What would the resulting binding look like..  x:1, y:1, z:1?  What would the Undo give us?  x:y, y:z, x:z  ?  If we're transitively closing the bindings on each lookup, then I'd think we'd have x:y, y:z, z:1 with an undo of x:y, y:z.  That seems like it should work, as long as the namespaces of the variables are always distinct.  Does this query work?  Query: p[z] = z.  Rule p[x] = y :- q[z] = (x,y).  Or does the z in the body of the rule get mistakenly bound to the z from the query?

An alternative, which seems like it will always give us the right answer, is where the datastructure is a linked-list of dictionaries.  On each bind we allocate another dictionary that contains the new bindings and push it onto the front of the list.  On each unbind, we pop the front element off the list.  On each lookup we traverse the list from front-to-back looking for a binding (recursively).  When binding a variable to a variable, we end up using a pointer to the variable we're binding to, which ensures the z's in the example above are different.