@timothyhinrichs 2016-07-11
Finished comments on this PR