@timothyhinrichs 2016-07-11
Forget about my binding comment.  Was confused about the roles of ctxt.Local versus ctxt. The reason is that there are 2 different ways that bindings get "undone" here.  One is the new one that backs up changes to a single ctxt.Local in the case of multiple rules with the same head.  The other one existed long before this patch and happens naturally when a recursive call to TopDown finishes (since then all the bindings made during the recursive call are "undone" by that ctxt being garbage-collected).   Anyway, LGTM once the constant-hash is fixed up.